updated foreign key check logic in original file
